refactor(discord-client): move event loop to a new file discord-loop.c
